+++ .github/actions/run-tests/action.yml
@@ -249,7 +249,7 @@ runs:
             fi
           fi
 
-          # Copy comparison images (saved by test_rendering_correctness.py).
+          # Copy comparison images (saved by source/isaaclab_tasks/test/test_rendering_*.py).
           local img_dir="$reports_dir/comparison-images"
           if [ -n "$volume_mount_source" ] && [ -d "${volume_mount_source}/tests/comparison-images" ]; then
             cp -r "${volume_mount_source}/tests/comparison-images" "$img_dir"
