+++ source/isaaclab_tasks/test/test_rendering_dexsuite_kuka.py
@@ -1,0 +1,36 @@
+# Copyright (c) 2022-2026, The Isaac Lab Project Developers (https://github.com/isaac-sim/IsaacLab/blob/main/CONTRIBUTORS.md).
+# All rights reserved.
+#
+# SPDX-License-Identifier: BSD-3-Clause
+
+"""Rendering correctness tests for Dexsuite Kuka-Allegro Lift backend combinations."""
+
+# Launch Isaac Sim Simulator first for kit-based combinations.
+from isaaclab.app import AppLauncher
+
+app_launcher = AppLauncher(headless=True, enable_cameras=True)
+simulation_app = app_launcher.app
+
+from pathlib import Path  # noqa: E402
+
+import pytest  # noqa: E402
+from rendering_test_utils import (  # noqa: E402
+    PHYSICS_RENDERER_AOV_COMBINATIONS,
+    make_attach_comparison_properties_fixture,
+    make_determinism_fixture,
+    make_generate_html_report_fixture,
+    rendering_test_dexsuite_kuka,
+)
+
+_COMPARISON_SCORES: list[dict] = []
+
+_determinism_fixture = make_determinism_fixture()
+_generate_html_report_fixture = make_generate_html_report_fixture(_COMPARISON_SCORES, Path(__file__).stem + ".html")
+_attach_comparison_properties_fixture = make_attach_comparison_properties_fixture(_COMPARISON_SCORES)
+
+
+@pytest.mark.flaky(max_runs=3, min_passes=1)
+@pytest.mark.parametrize("physics_backend,renderer,data_type", PHYSICS_RENDERER_AOV_COMBINATIONS)
+def test_rendering_dexsuite_kuka(physics_backend, renderer, data_type):
+    """Test dexsuite kuka allegro lift environment rendering correctness."""
+    rendering_test_dexsuite_kuka(physics_backend, renderer, data_type, _COMPARISON_SCORES)
