+++ source/isaaclab_tasks/test/test_rendering_shadow_hand_kitless.py
@@ -1,0 +1,31 @@
+# Copyright (c) 2022-2026, The Isaac Lab Project Developers (https://github.com/isaac-sim/IsaacLab/blob/main/CONTRIBUTORS.md).
+# All rights reserved.
+#
+# SPDX-License-Identifier: BSD-3-Clause
+
+"""Kit-less rendering correctness tests for Shadow Hand environment backend combinations."""
+
+from pathlib import Path
+
+import pytest
+from rendering_test_utils import (
+    KITLESS_PHYSICS_RENDERER_AOV_COMBINATIONS,
+    make_attach_comparison_properties_fixture,
+    make_determinism_fixture,
+    make_generate_html_report_fixture,
+    make_require_ovrtx_install_fixture,
+    rendering_test_shadow_hand,
+)
+
+_COMPARISON_SCORES: list[dict] = []
+
+_determinism_fixture = make_determinism_fixture()
+_generate_html_report_fixture = make_generate_html_report_fixture(_COMPARISON_SCORES, Path(__file__).stem + ".html")
+_attach_comparison_properties_fixture = make_attach_comparison_properties_fixture(_COMPARISON_SCORES)
+_require_ovrtx_install_fixture = make_require_ovrtx_install_fixture()
+
+
+@pytest.mark.parametrize("physics_backend,renderer,data_type", KITLESS_PHYSICS_RENDERER_AOV_COMBINATIONS)
+def test_rendering_shadow_hand_kitless(physics_backend, renderer, data_type):
+    """Test shadow hand environment rendering correctness."""
+    rendering_test_shadow_hand(physics_backend, renderer, data_type, _COMPARISON_SCORES)
